Split NetworkAppController into separate file
